requirements.txt
Django==1.8.1
boto==2.38.0
django-filter==0.9.2
django-bootstrap3==5.4.0
djangorestframework==3.1.1
django-cors-headers==1.0.0
requests==2.6.0
pytz==2015.2
mysqlclient==1.3.12
django-mysql==2.2.0
